+++ src/test/java/usePostgres/controller/FacultyControllerMVCTest.java
@@ -1,7 +1,9 @@
 package usePostgres.controller;
 
+import com.fasterxml.jackson.annotation.JacksonInject;
 import org.json.JSONObject;
 import org.junit.jupiter.api.Test;
+import org.mockito.InjectMocks;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -13,8 +15,9 @@ import usePostgres.models.Faculty;
 import usePostgres.repositories.FacultyRepository;
 import usePostgres.service.FacultyService;
 
-import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.when;
+import java.util.Optional;
+
+import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -28,25 +31,21 @@ public class FacultyControllerMVCTest {
     private MockMvc mockMvc;
     @MockBean
     private FacultyRepository facultyRepo;
-    @MockBean
-    private FacultyService facultyServ;
-
-    public static Faculty createFaculty() {
-
-        var faculty = new Faculty();
-        faculty.setId(1L);
-        faculty.setName("facultyTesting");
-        faculty.setColor("color");
-
-        return faculty;
-    }
 
     @Test
     public void add() throws Exception {
         var faculty = createFaculty();
-        var facultyJsonObj = creatJSONobjRecRequest(faculty);
+        faculty.setId(-1L);
 
-        when(facultyServ.add(any(Faculty.class))).thenReturn(faculty);
+        var facultyRes = createFaculty();
+
+        var facultyJsonObj = creatJSONobjRecRequest(faculty);
+        facultyJsonObj.put("id", -1L);
+        boolean existFaculty = true;
+
+        when(facultyRepo.existDataForFacultyName(any(String.class))).thenReturn(existFaculty);
+        when(facultyRepo.getMaxIdForTesting()).thenReturn(0L);
+        when(facultyRepo.save(any(Faculty.class))).thenReturn(facultyRes);
 
         mockMvc.perform(MockMvcRequestBuilders
                         .post("/faculty/add")
@@ -54,15 +53,15 @@ public class FacultyControllerMVCTest {
                         .contentType(MediaType.APPLICATION_JSON)
                         .accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.id").value(faculty.getId()))
-                .andExpect(jsonPath("$.name").value(faculty.getName()));
+                .andExpect(jsonPath("$.id").value(facultyRes.getId()))
+                .andExpect(jsonPath("$.name").value(facultyRes.getName()));
     }
 
     @Test
     public void read() throws Exception {
         var faculty = createFaculty();
 
-        when(facultyServ.read(any(Long.class))).thenReturn(faculty);
+        when(facultyRepo.findById(any(Long.class))).thenReturn(Optional.of(faculty));
 
         mockMvc.perform(MockMvcRequestBuilders
                         .get("/faculty/read/" + faculty.getId() )
@@ -75,7 +74,8 @@ public class FacultyControllerMVCTest {
     public void update() throws Exception {
         var faculty = createFaculty();
         var studentObj = creatJSONobjRecRequest(faculty);
-        when(facultyServ.update(any(Faculty.class))).thenReturn(faculty);
+
+        when(facultyRepo.save(any(Faculty.class))).thenReturn(faculty);
 
         mockMvc.perform(MockMvcRequestBuilders
                         .put("/faculty/update")
@@ -91,7 +91,8 @@ public class FacultyControllerMVCTest {
     public void delete() throws Exception {
         var faculty = createFaculty();
 
-        when(facultyServ.delete(any(Long.class))).thenReturn(faculty);
+        doNothing().when(facultyRepo).deleteById(any(Long.class));
+        when(facultyRepo.findById(any(Long.class))).thenReturn(Optional.of(faculty));
 
         mockMvc.perform(MockMvcRequestBuilders
                         .delete("/faculty/delete/"+ faculty.getId())
